+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+* migrate to null safety
+* migrate to android plugin v2
+
 ## 1.1.0
 
 * Add Mac OS support